refactor: extract AbstractInOutIterator to separate file
